requirements.txt
langchain-openai
streamlit
python-dotenv
httpx
google-generativeai